# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Dokumentasi API untuk layanan MongoDB pada aplikasi Go Fiber.",
        "title": "Go Fiber Mongo API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/go-fiber-mongo",
    "paths": {
...
        "/alumni/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil data alumni berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alumni (Mongo)"
                ],
                "summary": "Detail alumni",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Alumni",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mongo.GetAlumniByIDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mongo.GetAlumniByIDResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/mongo.GetAlumniByIDResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Memperbarui data alumni berdasarkan ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alumni (Mongo)"
                ],
                "summary": "Perbarui alumni",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Alumni",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Permintaan pembaruan alumni",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mongo.UpdateAlumniRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mongo.UpdateAlumniResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mongo.UpdateAlumniResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/mongo.UpdateAlumniResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus data alumni berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alumni (Mongo)"
                ],
                "summary": "Hapus alumni",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Alumni",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mongo.DeleteAlumniResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mongo.DeleteAlumniResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/mongo.DeleteAlumniResponse"
                        }
                    }
                }
            }
        },
...
        "/pekerjaan": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil daftar pekerjaan alumni dengan pagination, sorting, dan pencarian",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pekerjaan (Mongo)"
                ],
                "summary": "Daftar pekerjaan alumni",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Halaman",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Jumlah per halaman",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Kolom sortir (id, nama_perusahaan, posisi_jabatan, bidang_industri, lokasi_kerja, tanggal_mulai_kerja, status_pekerjaan, created_at)",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Urutan asc/desc",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Kata kunci pencarian",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mongo.GetAllPekerjaanResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Membuat data pekerjaan baru untuk alumni",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pekerjaan (Mongo)"
                ],
                "summary": "Tambah pekerjaan alumni",
                "parameters": [
                    {
                        "description": "Data pekerjaan alumni",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mongo.CreatePekerjaanAlumniRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/mongo.CreatePekerjaanAlumniResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mongo.CreatePekerjaanAlumniResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/mongo.CreatePekerjaanAlumniResponse"
                        }
                    }
                }
            }
        },
        "/pekerjaan/alumni/{alumni_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil seluruh pekerjaan milik alumni tertentu",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pekerjaan (Mongo)"
                ],
                "summary": "Daftar pekerjaan per alumni",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Alumni",
                        "name": "alumni_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mongo.GetPekerjaanAlumniByAlumniIDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mongo.GetPekerjaanAlumniByAlumniIDResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/mongo.GetPekerjaanAlumniByAlumniIDResponse"
                        }
                    }
                }
            }
        },
        "/pekerjaan/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil detail pekerjaan alumni berdasarkan ID pekerjaan",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pekerjaan (Mongo)"
                ],
                "summary": "Detail pekerjaan alumni",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Pekerjaan",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mongo.GetPekerjaanAlumniByIDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mongo.GetPekerjaanAlumniByIDResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/mongo.GetPekerjaanAlumniByIDResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Memperbarui data pekerjaan alumni berdasarkan ID pekerjaan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pekerjaan (Mongo)"
                ],
                "summary": "Perbarui pekerjaan alumni",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Pekerjaan",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data pekerjaan alumni yang diperbarui",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mongo.UpdatePekerjaanAlumniRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mongo.UpdatePekerjaanAlumniResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mongo.UpdatePekerjaanAlumniResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/mongo.UpdatePekerjaanAlumniResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus data pekerjaan alumni berdasarkan ID pekerjaan",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pekerjaan (Mongo)"
                ],
                "summary": "Hapus pekerjaan alumni",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Pekerjaan",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "fiber.Map": {
            "type": "object",
            "additionalProperties": true
        },
        "mongo.Alumni": {
            "type": "object",
            "properties": {
                "alamat": {
                    "type": "string"
                },
                "angkatan": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "jurusan": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "nim": {
                    "type": "string"
                },
                "no_telepon": {
                    "type": "string"
                },
                "role_id": {
                    "type": "string"
                },
                "tahun_lulus": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "mongo.CreatePekerjaanAlumniRequest": {
            "type": "object",
            "required": [
                "alumni_id",
                "bidang_industri",
                "lokasi_kerja",
                "nama_perusahaan",
                "posisi_jabatan",
                "status_pekerjaan",
                "tanggal_mulai_kerja"
            ],
            "properties": {
                "alumni_id": {
                    "type": "string"
                },
                "bidang_industri": {
                    "type": "string"
                },
                "deskripsi_pekerjaan": {
                    "type": "string"
                },
                "gaji_range": {
                    "type": "string"
                },
                "lokasi_kerja": {
                    "type": "string"
                },
                "nama_perusahaan": {
                    "type": "string"
                },
                "posisi_jabatan": {
                    "type": "string"
                },
                "status_pekerjaan": {
                    "type": "string",
                    "enum": [
                        "aktif",
                        "selesai",
                        "resigned"
                    ]
                },
                "tanggal_mulai_kerja": {
                    "type": "string"
                },
                "tanggal_selesai_kerja": {
                    "type": "string"
                }
            }
        },
        "mongo.CreatePekerjaanAlumniResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/mongo.PekerjaanAlumni"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "mongo.DeleteAlumniResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "mongo.GetAllPekerjaanResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/mongo.PekerjaanData"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "mongo.GetAlumniByIDResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/mongo.Alumni"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "mongo.GetPekerjaanAlumniByAlumniIDResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mongo.PekerjaanAlumni"
                    }
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "mongo.GetPekerjaanAlumniByIDResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/mongo.PekerjaanAlumni"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "mongo.MetaInfo": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "order": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "pages": {
                    "type": "integer"
                },
                "search": {
                    "type": "string"
                },
                "sortBy": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "mongo.PekerjaanAlumni": {
            "type": "object",
            "properties": {
                "alumni_id": {
                    "type": "string"
                },
                "bidang_industri": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deskripsi_pekerjaan": {
                    "type": "string"
                },
                "gaji_range": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lokasi_kerja": {
                    "type": "string"
                },
                "nama_perusahaan": {
                    "type": "string"
                },
                "posisi_jabatan": {
                    "type": "string"
                },
                "status_pekerjaan": {
                    "type": "string"
                },
                "tanggal_mulai_kerja": {
                    "type": "string"
                },
                "tanggal_selesai_kerja": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "mongo.PekerjaanData": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mongo.PekerjaanAlumni"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/mongo.MetaInfo"
                }
            }
        },
...
        "mongo.UpdateAlumniRequest": {
            "type": "object",
            "properties": {
                "alamat": {
                    "type": "string"
                },
                "angkatan": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "jurusan": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "nim": {
                    "type": "string"
                },
                "no_telepon": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role_id": {
                    "type": "string"
                },
                "tahun_lulus": {
                    "type": "integer"
                }
            }
        },
        "mongo.UpdateAlumniResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/mongo.Alumni"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "mongo.UpdatePekerjaanAlumniRequest": {
            "type": "object",
            "required": [
                "bidang_industri",
                "lokasi_kerja",
                "nama_perusahaan",
                "posisi_jabatan",
                "status_pekerjaan",
                "tanggal_mulai_kerja"
            ],
            "properties": {
                "bidang_industri": {
                    "type": "string"
                },
                "deskripsi_pekerjaan": {
                    "type": "string"
                },
                "gaji_range": {
                    "type": "string"
                },
                "lokasi_kerja": {
                    "type": "string"
                },
                "nama_perusahaan": {
                    "type": "string"
                },
                "posisi_jabatan": {
                    "type": "string"
                },
                "status_pekerjaan": {
                    "type": "string",
                    "enum": [
                        "aktif",
                        "selesai",
                        "resigned"
                    ]
                },
                "tanggal_mulai_kerja": {
                    "type": "string"
                },
                "tanggal_selesai_kerja": {
                    "type": "string"
                }
            }
        },
        "mongo.UpdatePekerjaanAlumniResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/mongo.PekerjaanAlumni"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}